refactor: Moved AudioPlayer to examples and remove sounddevice direct dep. Thanks @domstoppable
