+++ pyproject.toml
@@ -28,7 +28,6 @@ dependencies = [
     "pupil-labs-neon-recording>=1.0.1",
     "pupil-labs-video>=1.0.3",
     "pydantic>=2.10.6",
-    "sounddevice>=0.5.2",
     "websockets>=15.0.1",
     "zeroconf>=0.146.1",
 ]
@@ -69,7 +68,7 @@ dev = [
     "tomli>=2.0; python_version < '3.11'",
     "types-deprecated>=1.2.15.20250304",
 ]
-examples = ["opencv-python", "rich", "librosa", "speechrecognition", "openai-whisper"]
+examples = ["opencv-python", "sounddevice", "rich", "librosa", "speechrecognition", "openai-whisper"]
 
 [build-system]
 requires = ["hatchling", "uv-dynamic-versioning"]
@@ -87,10 +86,10 @@ branch = true
 source = ["pupil_labs"]
 
 [tool.deptry.per_rule_ignores]
-DEP001 = ["pupil_labs", "cv2", "whisper", "speech_recognition"]
+DEP001 = ["pupil_labs", "cv2", "whisper", "speech_recognition", "sounddevice"]
 DEP002 = ["opencv-python"]
 DEP003 = ["pupil_labs", "typing_extensions", "pydantic_core"]
-DEP004 = ["cv2", "mkdocs", "rich", "librosa", "whisper", "speech_recognition"]
+DEP004 = ["cv2", "mkdocs", "rich", "librosa", "whisper", "speech_recognition", "sounddevice"]
 
 [tool.hatch.version]
 source = "uv-dynamic-versioning"
