+++ src/pupil_labs/realtime_api/__init__.py
@@ -9,7 +9,6 @@ import importlib.metadata
 
 from .device import APIPath, Device, DeviceError, StatusUpdateNotifier
 from .discovery import Network, discover_devices
-from .models import AudioPlayer
 from .streaming import (
     AudioFrame,
     BlinkEventData,
@@ -42,7 +41,6 @@ except importlib.metadata.PackageNotFoundError:
 __all__ = [
     "APIPath",
     "AudioFrame",
-    "AudioPlayer",
     "BlinkEventData",
     "Device",
     "DeviceError",
